Add files via upload
So far, I've created a login screen where a user can either input their already-existing username and password, or create an account and then log in. The application quits after logging in since I haven't created the UI for the music library yet. As a backdoor, the username/password Admin/123456 is hard-coded in case we somehow lose all other accounts.

Next steps:
1. Mask password with asterisks (123456 = ******)
2. Create the UI for the music library